updates segmentation tests

+++ tests/experiment_test.py
@@ -1,7 +1,9 @@
 import os
+import glob
 import shutil
 import pytest
 import json
+import pandas as pd
 from fuse_toolkit import Experiment
 
 class TestExperimentClass:
@@ -13,14 +15,14 @@ class TestExperimentClass:
         self.test_folder = os.path.join('tests', 'test_resources', "")
         self.date = '2023-12-31'
         self.expID = 'TestExp'
-        self.path = os.path.join('tests', 'test_resources', 'image.tif')
+        self.path = os.path.join('tests', 'test_resources', 'Tester_01.tif')
         self.parseID = 'Name_Well'
         self.separator = '_'
         self.channelInfo = 'RFP_GFP'
         self.channelToSegment = 'RFP'
         self.numFrames = 90
         self.frameInterval = 1
-        self.frameToSegment = 'all'
+        self.frameToSegment = 0
         self.expNote = 'test note'
         self.test_folder = os.path.join('tests', 'test_resources',
                                         self.date + "_" + self.expID)
@@ -28,9 +30,9 @@ class TestExperimentClass:
         self.test_json_path = os.path.join('tests', 'test_resources',
                                            self.json_date + "_" + self.expID,
                                            'info_exp.json')
-
-        yield
-        shutil.rmtree(self.test_folder)
+        yield  
+        if os.path.isdir(self.test_folder):
+            shutil.rmtree(self.test_folder)
 
         
     def test_initialization(self, setup):
@@ -48,9 +50,15 @@ class TestExperimentClass:
         assert experiment.channelToSegment == self.channelToSegment
         assert experiment.numFrames == self.numFrames
         assert experiment.frameInterval == self.frameInterval
-        assert experiment.frameToSegment == 'all'
+        assert experiment.frameToSegment == self.frameToSegment
         assert experiment.expNote == self.expNote
         assert hasattr(experiment, 'folder')
+        assert hasattr(experiment, 'df')
+        if os.path.exists(os.path.join(experiment.folder, f"{experiment.date}_{experiment.expID}", f"{experiment.date}_{experiment.expID}.csv")):
+            assert isinstance(experiment.df, pd.DataFrame)
+        else:
+            assert experiment.df is None
+
 
     def test_folder_creation(self, setup):
         # Test if the folder and its subdirectories are created correctly
@@ -70,6 +78,7 @@ class TestExperimentClass:
         # Info file
         info_file_path = os.path.join(exp_folder_path, 'info_exp.json')
         assert os.path.exists(info_file_path), "Experiment info file not created correctly."
+
 
     def test_info_file_creation(self, setup):
         # Test if the info_exp.json file is created correctly
@@ -104,6 +113,7 @@ class TestExperimentClass:
         assert info_data['frameToSegment'] == expected_frameToSegment
         assert info_data['expNote'] == self.expNote
 
+
     def test_from_json(self, setup):
         # Test if the Experiment object is initialized correctly from JSON
         experiment = Experiment.from_json(self.test_json_path)
@@ -120,24 +130,39 @@ class TestExperimentClass:
         assert experiment.frameToSegment == self.frameToSegment
         assert experiment.expNote == self.expNote
         assert experiment.multiChannel == (len(self.channelInfo.split(self.separator)) > 1)
-        
+
+     
     def test_preview_segmentation(self, setup):
-        # Initialize the Experiment object with test data
+        # Test if preview_segmentation() without errors
         experiment = Experiment(self.date, self.expID, self.path, self.parseID,
                                 self.separator, self.channelInfo, self.channelToSegment,
                                 self.numFrames, self.frameInterval, self.frameToSegment,
                                 self.expNote)
 
-        # Test parameters for preview_segmentation
-        min_size = 100
-        flow_threshold = 0.4
-        mask_threshold = 0.0
-        rescale = 1
-        diameter = 30
+        experiment.preview_segmentation(output=False)
 
-        try:
-            experiment.preview_segmentation(min_size, flow_threshold, mask_threshold,
-                                            rescale, diameter, output=False)
-            assert True
-        except Exception as e:
-            pytest.fail(f"preview_segmentation method failed: {e}")
+
+    def test_cell_segmentation(self, setup):
+        # Tests that cell_segmentation() runs without errors and validates results
+        experiment = Experiment(self.date, self.expID, self.path, self.parseID,
+                                self.separator, self.channelInfo, self.channelToSegment,
+                                self.numFrames, self.frameInterval, self.frameToSegment,
+                                self.expNote)
+
+        result_df = experiment.segment_cells()
+
+        assert isinstance(result_df, pd.DataFrame)
+        assert not result_df.empty
+        assert isinstance(experiment.df, pd.DataFrame)
+        assert experiment.df.equals(result_df)
+        
+        if os.path.isdir(experiment.path):
+            input_files = sorted(glob.glob(os.path.join(experiment.path, '*.tif')))
+        elif os.path.splitext(experiment.path)[-1].lower() == '.tif':
+            input_files = [experiment.path]
+
+        for input_file in input_files:
+            base_name = os.path.basename(input_file).split('.')[0]
+            expected_seg_file = os.path.join(experiment.folder, f"{experiment.date}_{experiment.expID}", 
+                                             "segmentations", f"{base_name}_seg.tif")
+            assert os.path.isfile(expected_seg_file), f"Segmentation file not found for {base_name}"
